# api/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Tekton Hub
  description: HTTP service for managing Tekton Hub
  version: ""
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /categories:
    get:
      tags:
      - category
      summary: Categories category
      description: Get all Categories with their associated tags.
      operationId: category#Categories
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/CategoryCategoriesResponseBody'
      schemes:
      - http
  /resource/{resourceID}/rating:
    get:
      tags:
      - resource
      summary: GetRating resource
      description: Get User's rating of a resource
      operationId: resource#GetRating
      parameters:
      - name: resourceID
        in: path
        description: Id of the Resource
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ResourceGetRatingResponseBody'
            required:
            - rating
      schemes:
      - http
    put:
      tags:
      - resource
      summary: UpdateRating resource
      description: Update User's rating of a resource
      operationId: resource#UpdateRating
      parameters:
      - name: resourceID
        in: path
        description: Id of the Resource
        required: true
        type: integer
      - name: UpdateRatingRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/ResourceUpdateRatingRequestBody'
          required:
          - rating
      responses:
        "200":
          description: OK response.
      schemes:
      - http
  /resource/{resourceID}/versions:
    get:
      tags:
      - resource
      summary: AllVersions resource
      description: Get all versions of a resource
      operationId: resource#AllVersions
      parameters:
      - name: resourceID
        in: path
        description: Id of the Resource
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              $ref: '#/definitions/ResourceResponse'
      schemes:
      - http
  /resources:
    get:
      tags:
      - resource
      summary: All resource
      description: Get all tasks and pipelines.
      operationId: resource#All
      parameters:
      - name: limit
        in: query
        description: Number of Resource to fetch
        required: false
        type: integer
        default: 100
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ResourceResourceResponseCollection'
      schemes:
      - http
  /swagger/swagger.json:
    get:
      tags:
      - swagger
      summary: Download gen/http/openapi.json
      description: JSON document containing the API swagger definition
      operationId: swagger#/swagger/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
definitions:
  CatalogResponse:
    title: CatalogResponse
    type: object
    properties:
      id:
        type: integer
        description: ID is the unique id of the catalog
        example: 1
        format: int64
      type:
        type: string
        description: Type of catalog
        example: Official
    example:
      id: 1
      type: Official
    required:
    - id
    - type
  CategoryCategoriesResponseBody:
    title: CategoryCategoriesResponseBody
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/CategoryResponseBody'
        example:
        - id: 1
          name: Notification
          tags:
          - id: 1
            name: notification
          - id: 1
            name: notification
          - id: 1
            name: notification
          - id: 1
            name: notification
        - id: 1
          name: Notification
          tags:
          - id: 1
            name: notification
          - id: 1
            name: notification
          - id: 1
            name: notification
          - id: 1
            name: notification
        - id: 1
          name: Notification
          tags:
          - id: 1
            name: notification
          - id: 1
            name: notification
          - id: 1
            name: notification
          - id: 1
            name: notification
        - id: 1
          name: Notification
          tags:
          - id: 1
            name: notification
          - id: 1
            name: notification
          - id: 1
            name: notification
          - id: 1
            name: notification
      errors:
        type: array
        items:
          type: string
          example: Nesciunt placeat provident ducimus amet doloribus molestias.
        example:
        - Non omnis quas deserunt.
        - Assumenda pariatur.
    example:
      data:
      - id: 1
        name: Notification
        tags:
        - id: 1
          name: notification
        - id: 1
          name: notification
        - id: 1
          name: notification
        - id: 1
          name: notification
      - id: 1
        name: Notification
        tags:
        - id: 1
          name: notification
        - id: 1
          name: notification
        - id: 1
          name: notification
        - id: 1
          name: notification
      errors:
      - Maiores quaerat consequatur ullam quia nihil officia.
      - Non aut qui dolor.
  CategoryResponseBody:
    title: CategoryResponseBody
    type: object
    properties:
      id:
        type: integer
        description: ID is the unique id of the category
        example: 1
        format: int64
      name:
        type: string
        description: Name of the Category
        example: Notification
      tags:
        type: array
        items:
          $ref: '#/definitions/ResourceTagResponseBody'
        description: Tags associated with the category
        example:
        - id: 1
          name: notification
        - id: 1
          name: notification
        - id: 1
          name: notification
    example:
      id: 1
      name: Notification
      tags:
      - id: 1
        name: notification
      - id: 1
        name: notification
      - id: 1
        name: notification
      - id: 1
        name: notification
    required:
    - id
    - name
    - tags
  ResourceGetRatingResponseBody:
    title: ResourceGetRatingResponseBody
    type: object
    properties:
      rating:
        type: integer
        description: Rating of the resource
        example: 3
        minimum: 0
        maximum: 5
    example:
      rating: 3
    required:
    - rating
  ResourceResourceResponseCollection:
    title: 'Mediatype identifier: application/vnd.hub.resource; type=collection; view=default'
    type: array
    items:
      $ref: '#/definitions/ResourceResponse'
    description: AllResponseBody is the result type for an array of ResourceResponse
      (default view)
    example:
    - catalog:
        id: 1
        type: Official
      description: Buildah is a build tool to create images from source code.
      id: 1
      last_updated_at: "1232342423"
      name: buildah
      rating: 2.3
      tags:
      - id: 1
        name: notification
      - id: 1
        name: notification
      type: Task
      versions:
      - id: 1
        version: "2.3"
      - id: 1
        version: "2.3"
      - id: 1
        version: "2.3"
    - catalog:
        id: 1
        type: Official
      description: Buildah is a build tool to create images from source code.
      id: 1
      last_updated_at: "1232342423"
      name: buildah
      rating: 2.3
      tags:
      - id: 1
        name: notification
      - id: 1
        name: notification
      type: Task
      versions:
      - id: 1
        version: "2.3"
      - id: 1
        version: "2.3"
      - id: 1
        version: "2.3"
  ResourceResponse:
    title: 'Mediatype identifier: application/vnd.hub.resource; view=default'
    type: object
    properties:
      catalog:
        $ref: '#/definitions/CatalogResponse'
      description:
        type: string
        description: Description of the resource
        example: Buildah is a build tool to create images from source code.
      id:
        type: integer
        description: ID is the unique id of the resource
        example: 1
        format: int64
      last_updated_at:
        type: string
        description: TimeStamp the resource last updated at
        example: "1232342423"
      name:
        type: string
        description: Name of the resource
        example: buildah
      rating:
        type: number
        description: Rating of resource
        example: 2.3
        minimum: 0
        maximum: 5
      tags:
        type: array
        items:
          $ref: '#/definitions/ResourceTagResponse'
        description: Tags associated to the resource
        example:
        - id: 1
          name: notification
        - id: 1
          name: notification
      type:
        type: string
        description: Type of the resource
        example: Task
      versions:
        type: array
        items:
          $ref: '#/definitions/ResourceVersionResponse'
        description: Different Versions of the resource
        example:
        - id: 1
          version: "2.3"
        - id: 1
          version: "2.3"
    description: Describes a resource retrieved from catalog (default view)
    example:
      catalog:
        id: 1
        type: Official
      description: Buildah is a build tool to create images from source code.
      id: 1
      last_updated_at: "1232342423"
      name: buildah
      rating: 2.3
      tags:
      - id: 1
        name: notification
      - id: 1
        name: notification
      type: Task
      versions:
      - id: 1
        version: "2.3"
      - id: 1
        version: "2.3"
      - id: 1
        version: "2.3"
    required:
    - id
    - name
    - catalog
    - description
    - type
    - versions
    - tags
    - rating
    - last_updated_at
  ResourceTagResponse:
    title: ResourceTagResponse
    type: object
    properties:
      id:
        type: integer
        description: ID is the unique id of the tag
        example: 1
        format: int64
      name:
        type: string
        description: Name of the tag
        example: notification
    example:
      id: 1
      name: notification
    required:
    - id
    - name
  ResourceTagResponseBody:
    title: ResourceTagResponseBody
    type: object
    properties:
      id:
        type: integer
        description: ID is the unique id of the tag
        example: 1
        format: int64
      name:
        type: string
        description: Name of the tag
        example: notification
    example:
      id: 1
      name: notification
    required:
    - id
    - name
  ResourceUpdateRatingRequestBody:
    title: ResourceUpdateRatingRequestBody
    type: object
    properties:
      rating:
        type: integer
        description: Rating of resource to be updated
        example: 12809560377624145324
        format: int64
    example:
      rating: 2349809637697434414
    required:
    - rating
  ResourceVersionResponse:
    title: ResourceVersionResponse
    type: object
    properties:
      id:
        type: integer
        description: ID is the unique id of the version
        example: 1
        format: int64
      version:
        type: string
        description: Version of resource
        example: "2.3"
    example:
      id: 1
      version: "2.3"
    required:
    - id
    - version